# openwebui-docs/docs.openwebui.com/mcp-agent/bridge_openapi.json
{"openapi":"3.1.0","info":{"title":"MCP Bridge Server","description":"Bridge server that exposes MCP (Model Context Protocol) functionality via OpenAPI endpoints","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Root","description":"Root endpoint with server information","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Root  Get"}}}}}}},"/execute_cua_command":{"post":{"summary":"Execute a command on the CUA MCP server","description":"Executes a command on the Computer Use Agent (CUA) MCP server and returns the result. This endpoint bridges OpenWebUI's stateless OpenAPI calls to the stateful MCP protocol.","operationId":"execute_cua_command_execute_cua_command_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommandRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommandResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/list_desktop_files":{"post":{"summary":"List files on the desktop","description":"Convenience endpoint to list files on the desktop. This is a specialized wrapper around the execute_cua_command endpoint.","operationId":"list_desktop_files_list_desktop_files_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommandResponse"}}}}}}},"/read_desktop_file/{filename}":{"post":{"summary":"Read a file from the desktop","description":"Read the contents of a specific file from the desktop directory.","operationId":"read_desktop_file_read_desktop_file__filename__post","parameters":[{"name":"filename","in":"path","required":true,"schema":{"type":"string","title":"Filename"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommandResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp_state":{"get":{"summary":"Get current MCP server state","description":"Retrieves the current state of the MCP server connection, including session information and command history.","operationId":"get_mcp_state_mcp_state_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StateResponse"}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Health Check Health Get"}}}}}}},"/available_commands":{"get":{"summary":"Get list of available MCP commands","description":"Returns a list of commands that can be executed on the MCP server.","operationId":"get_available_commands_available_commands_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Get Available Commands Available Commands Get"}}}}}}}},"components":{"schemas":{"CommandRequest":{"properties":{"command":{"type":"string","title":"Command","description":"The command to execute on the MCP server","example":"list files on the desktop"}},"type":"object","required":["command"],"title":"CommandRequest","description":"Request model for executing MCP commands"},"CommandResponse":{"properties":{"status":{"type":"string","title":"Status","description":"Execution status (success or error)"},"command":{"type":"string","title":"Command","description":"The command that was executed"},"result":{"additionalProperties":true,"type":"object","title":"Result","description":"Command execution result"},"observation":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Observation","description":"Human-readable observation from the MCP server"}},"type":"object","required":["status","command","result"],"title":"CommandResponse","description":"Response model for command execution results"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"StateResponse":{"properties":{"connected":{"type":"boolean","title":"Connected","description":"Whether the MCP client is connected"},"session_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Session Id","description":"Current session ID"},"current_directory":{"type":"string","title":"Current Directory","description":"Current working directory"},"command_count":{"type":"integer","title":"Command Count","description":"Number of commands executed in this session"}},"type":"object","required":["connected","session_id","current_directory","command_count"],"title":"StateResponse","description":"Response model for MCP server state"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}